+++ resources/js/pages/Files/Browser.vue
@@ -7,9 +7,27 @@
                     id="back2site" content="View Application" v-if="site"
                     :data-tooltip="'Open overview for ' + site.name"
                     data-position="left center" />
+
+                <sui-button class="icon" floated="right" @click.native="toggleNewFile">
+                    <i class="icons">
+                        <sui-icon name="file outline" />
+                        <sui-icon name="add" class="purple corner" />
+                    </i>
+                </sui-button>
             </path-bar>
 
             <file-list :files="files" :path="currentPath" />
+
+            <sui-modal size="tiny" v-model="promptNewFile">
+                <sui-modal-header>Enter filename</sui-modal-header>
+                <sui-modal-content>
+                    <sui-input class="fluid" v-model="filename" @keyup.enter="edit"
+                        style="border: 1px solid rgba(34, 36, 38, 0.15)" />
+                </sui-modal-content>
+                <sui-modal-actions>
+                    <sui-button positive @click.native="edit" content="OK" />
+                </sui-modal-actions>
+            </sui-modal>
         </sui-grid-column>
     </sui-grid>
 </template>
@@ -28,6 +46,12 @@ export default {
         this.$store.dispatch('files/load', { path: to.params.path });
         next();
     },
+    data() {
+        return {
+            filename: '',
+            promptNewFile: false,
+        };
+    },
     props: [
         'path',
     ],
@@ -41,8 +65,22 @@ export default {
             findSite: 'sites/findByDocroot',
             files: 'files/all',
         }),
+        createPath() {
+            return `${this.currentPath}/${this.filename}`;
+        },
         site() {
             return this.findSite(this.currentPath);
+        },
+    },
+    methods: {
+        edit() {
+            this.$router.push({
+                name: 'files.edit',
+                query: { f: this.createPath },
+            });
+        },
+        toggleNewFile() {
+            this.promptNewFile = !this.promptNewFile;
         },
     },
 };